+++ src/main/kotlin/app/morphe/gui/di/AppModule.kt
@@ -1,3 +1,8 @@
+/*
+ * Copyright 2026 Morphe.
+ * https://github.com/MorpheApp/morphe-cli
+ */
+
 package app.morphe.gui.di
 
 import app.morphe.gui.data.repository.ConfigRepository