+++ src/main/kotlin/app/morphe/gui/ui/screens/home/HomeViewModel.kt
@@ -1,3 +1,8 @@
+/*
+ * Copyright 2026 Morphe.
+ * https://github.com/MorpheApp/morphe-cli
+ */
+
 package app.morphe.gui.ui.screens.home
 
 import cafe.adriel.voyager.core.model.ScreenModel