+++ CHANGELOG.md
@@ -1,3 +1,223 @@
+## [1.6.4-dev.1](https://github.com/MorpheApp/morphe-cli/compare/v1.6.3...v1.6.4-dev.1) (2026-03-30)
+
+
+### Bug Fixes
+
+* Use patcher implementation of strip libs ([#83](https://github.com/MorpheApp/morphe-cli/issues/83)) ([43f50ea](https://github.com/MorpheApp/morphe-cli/commit/43f50ea133088090648fa318047626f3166b8639)), closes [#80](https://github.com/MorpheApp/morphe-cli/issues/80) [#82](https://github.com/MorpheApp/morphe-cli/issues/82)
+
+## [1.6.3](https://github.com/MorpheApp/morphe-cli/compare/v1.6.2...v1.6.3) (2026-03-28)
+
+
+### Bug Fixes
+
+* Handle patching XAPK files ([#93](https://github.com/MorpheApp/morphe-cli/issues/93)) ([5d18490](https://github.com/MorpheApp/morphe-cli/commit/5d18490305f10949bf5407bc7167549e3a1cc8f4))
+* Handle running CLI in headless environment ([#95](https://github.com/MorpheApp/morphe-cli/issues/95)) ([70f2952](https://github.com/MorpheApp/morphe-cli/commit/70f2952b09de57ad7c06bd92365025d38164ed26))
+* Update to latest Patcher ([fbc6c9e](https://github.com/MorpheApp/morphe-cli/commit/fbc6c9e2dbab2db9b46f442feeec1a2df565114e))
+* Update to Patcher 1.3.3 ([ca579fb](https://github.com/MorpheApp/morphe-cli/commit/ca579fbc8afa32b98a77475ae4db0792d1165a3d))
+
+## [1.6.3-dev.4](https://github.com/MorpheApp/morphe-cli/compare/v1.6.3-dev.3...v1.6.3-dev.4) (2026-03-28)
+
+
+### Bug Fixes
+
+* Handle patching XAPK files ([#93](https://github.com/MorpheApp/morphe-cli/issues/93)) ([5d18490](https://github.com/MorpheApp/morphe-cli/commit/5d18490305f10949bf5407bc7167549e3a1cc8f4))
+
+## [1.6.3-dev.3](https://github.com/MorpheApp/morphe-cli/compare/v1.6.3-dev.2...v1.6.3-dev.3) (2026-03-28)
+
+
+### Bug Fixes
+
+* Handle running CLI in headless environment ([#95](https://github.com/MorpheApp/morphe-cli/issues/95)) ([70f2952](https://github.com/MorpheApp/morphe-cli/commit/70f2952b09de57ad7c06bd92365025d38164ed26))
+
+## [1.6.3-dev.2](https://github.com/MorpheApp/morphe-cli/compare/v1.6.3-dev.1...v1.6.3-dev.2) (2026-03-24)
+
+
+### Bug Fixes
+
+* Update to latest Patcher ([fbc6c9e](https://github.com/MorpheApp/morphe-cli/commit/fbc6c9e2dbab2db9b46f442feeec1a2df565114e))
+
+## [1.6.3-dev.1](https://github.com/MorpheApp/morphe-cli/compare/v1.6.2...v1.6.3-dev.1) (2026-03-23)
+
+
+### Bug Fixes
+
+* Update to Patcher 1.3.3 ([ca579fb](https://github.com/MorpheApp/morphe-cli/commit/ca579fbc8afa32b98a77475ae4db0792d1165a3d))
+
+## [1.6.2](https://github.com/MorpheApp/morphe-cli/compare/v1.6.1...v1.6.2) (2026-03-22)
+
+
+### Bug Fixes
+
+* Update to Patcher 1.3.2 ([f4c00bb](https://github.com/MorpheApp/morphe-cli/commit/f4c00bbea80997c816f40f5dcf0a17f771d80b43))
+
+## [1.6.2-dev.1](https://github.com/MorpheApp/morphe-cli/compare/v1.6.1...v1.6.2-dev.1) (2026-03-22)
+
+
+### Bug Fixes
+
+* Update to Patcher 1.3.2 ([f4c00bb](https://github.com/MorpheApp/morphe-cli/commit/f4c00bbea80997c816f40f5dcf0a17f771d80b43))
+
+## [1.6.1](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0...v1.6.1) (2026-03-22)
+
+
+### Bug Fixes
+
+* Update to Patcher 1.3.1 ([39bdd44](https://github.com/MorpheApp/morphe-cli/commit/39bdd44d13cf99b855d4e8501a47283f0681cbe7))
+
+## [1.6.1-dev.1](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0...v1.6.1-dev.1) (2026-03-22)
+
+
+### Bug Fixes
+
+* Update to Patcher 1.3.1 ([39bdd44](https://github.com/MorpheApp/morphe-cli/commit/39bdd44d13cf99b855d4e8501a47283f0681cbe7))
+
+# [1.6.0](https://github.com/MorpheApp/morphe-cli/compare/v1.5.0...v1.6.0) (2026-03-22)
+
+
+### Bug Fixes
+
+* Bump arsclib version, remove APKEditor ([412847e](https://github.com/MorpheApp/morphe-cli/commit/412847e5a0b990321b7443fe774f7420189035d6))
+* Dispatch error when launching app ([#79](https://github.com/MorpheApp/morphe-cli/issues/79)) ([989e72a](https://github.com/MorpheApp/morphe-cli/commit/989e72a04c9ea1b512d556ff8db48b99a4817d98))
+* Update to latest patcher dev ([496d8e9](https://github.com/MorpheApp/morphe-cli/commit/496d8e9d2d1e402f7246b6f3adb85c4c6c1865ca))
+* Update to latest patcher dev ([1356259](https://github.com/MorpheApp/morphe-cli/commit/135625917a54eb5e0a9464c40d0ce504619b4845))
+* Update to latest patcher dev ([fb8a4ff](https://github.com/MorpheApp/morphe-cli/commit/fb8a4ffe70ae24c74c497608b9eb9aa2d76f2a8a))
+* Update to latest patcher dev ([1551c8e](https://github.com/MorpheApp/morphe-cli/commit/1551c8ef81bd4d267833c1dc543d13c1d1def29f))
+* Update to Patcher 1.3.0 ([19c0846](https://github.com/MorpheApp/morphe-cli/commit/19c084612310cb590c90b809403450abf6a81560))
+* Use GitHub raw links for CLI update check ([#87](https://github.com/MorpheApp/morphe-cli/issues/87)) ([2aa6263](https://github.com/MorpheApp/morphe-cli/commit/2aa62637098fa52dcc79422116b91487aed5600a))
+* Use more logging verbosity when updating options.json ([#72](https://github.com/MorpheApp/morphe-cli/issues/72)) ([4d223f2](https://github.com/MorpheApp/morphe-cli/commit/4d223f299b690c100ff74ece7e4415f562c6ac96))
+
+
+### Features
+
+* Add desktop GUI ([#42](https://github.com/MorpheApp/morphe-cli/issues/42)) ([cd627e6](https://github.com/MorpheApp/morphe-cli/commit/cd627e6a4320084403beacd414278b3b75f469da))
+* Inform user if cli is outdated ([#81](https://github.com/MorpheApp/morphe-cli/issues/81)) ([5bcd4f7](https://github.com/MorpheApp/morphe-cli/commit/5bcd4f7c612ec3aad732e83ea131a019424b3bb8))
+
+# [1.6.0-dev.8](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0-dev.7...v1.6.0-dev.8) (2026-03-21)
+
+
+### Bug Fixes
+
+* Use GitHub raw links for CLI update check ([#87](https://github.com/MorpheApp/morphe-cli/issues/87)) ([2aa6263](https://github.com/MorpheApp/morphe-cli/commit/2aa62637098fa52dcc79422116b91487aed5600a))
+
+# [1.6.0-dev.7](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0-dev.6...v1.6.0-dev.7) (2026-03-20)
+
+
+### Bug Fixes
+
+* Update to latest patcher dev ([496d8e9](https://github.com/MorpheApp/morphe-cli/commit/496d8e9d2d1e402f7246b6f3adb85c4c6c1865ca))
+
+# [1.6.0-dev.6](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0-dev.5...v1.6.0-dev.6) (2026-03-20)
+
+
+### Features
+
+* Inform user if cli is outdated ([#81](https://github.com/MorpheApp/morphe-cli/issues/81)) ([5bcd4f7](https://github.com/MorpheApp/morphe-cli/commit/5bcd4f7c612ec3aad732e83ea131a019424b3bb8))
+
+# [1.6.0-dev.5](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0-dev.4...v1.6.0-dev.5) (2026-03-16)
+
+
+### Bug Fixes
+
+* Update to latest patcher dev ([1356259](https://github.com/MorpheApp/morphe-cli/commit/135625917a54eb5e0a9464c40d0ce504619b4845))
+
+# [1.6.0-dev.4](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0-dev.3...v1.6.0-dev.4) (2026-03-15)
+
+
+### Bug Fixes
+
+* Update to latest patcher dev ([fb8a4ff](https://github.com/MorpheApp/morphe-cli/commit/fb8a4ffe70ae24c74c497608b9eb9aa2d76f2a8a))
+
+# [1.6.0-dev.3](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0-dev.2...v1.6.0-dev.3) (2026-03-15)
+
+
+### Bug Fixes
+
+* Update to latest patcher dev ([1551c8e](https://github.com/MorpheApp/morphe-cli/commit/1551c8ef81bd4d267833c1dc543d13c1d1def29f))
+
+# [1.6.0-dev.2](https://github.com/MorpheApp/morphe-cli/compare/v1.6.0-dev.1...v1.6.0-dev.2) (2026-03-15)
+
+
+### Bug Fixes
+
+* Dispatch error when launching app ([#79](https://github.com/MorpheApp/morphe-cli/issues/79)) ([989e72a](https://github.com/MorpheApp/morphe-cli/commit/989e72a04c9ea1b512d556ff8db48b99a4817d98))
+
+# [1.6.0-dev.1](https://github.com/MorpheApp/morphe-cli/compare/v1.5.1-dev.3...v1.6.0-dev.1) (2026-03-14)
+
+
+### Features
+
+* Add desktop GUI ([#42](https://github.com/MorpheApp/morphe-cli/issues/42)) ([cd627e6](https://github.com/MorpheApp/morphe-cli/commit/cd627e6a4320084403beacd414278b3b75f469da))
+
+## [1.5.1-dev.3](https://github.com/MorpheApp/morphe-cli/compare/v1.5.1-dev.2...v1.5.1-dev.3) (2026-03-13)
+
+
+### Bug Fixes
+
+* Use more logging verbosity when updating options.json ([#72](https://github.com/MorpheApp/morphe-cli/issues/72)) ([4d223f2](https://github.com/MorpheApp/morphe-cli/commit/4d223f299b690c100ff74ece7e4415f562c6ac96))
+
+## [1.5.1-dev.2](https://github.com/MorpheApp/morphe-cli/compare/v1.5.1-dev.1...v1.5.1-dev.2) (2026-03-13)
+
+
+### Bug Fixes
+
+* Bump arsclib version, remove APKEditor ([412847e](https://github.com/MorpheApp/morphe-cli/commit/412847e5a0b990321b7443fe774f7420189035d6))
+
+## [1.5.1-dev.1](https://github.com/MorpheApp/morphe-cli/compare/v1.5.0...v1.5.1-dev.1) (2026-03-13)
+
+
+### Bug Fixes
+
+* Update to Patcher 1.3.0 ([19c0846](https://github.com/MorpheApp/morphe-cli/commit/19c084612310cb590c90b809403450abf6a81560))
+
+# [1.5.0](https://github.com/MorpheApp/morphe-cli/compare/v1.4.0...v1.5.0) (2026-03-07)
+
+
+### Bug Fixes
+
+* Use latest library dev release ([b60ea86](https://github.com/MorpheApp/morphe-cli/commit/b60ea86c72b8a62e2220b3a76dba01d47c494750))
+* Use latest patcher and library dev release ([2f21880](https://github.com/MorpheApp/morphe-cli/commit/2f21880c49705b3e5153b7e636f01cf578d7b1c0))
+* Use latest patcher dev release ([ea6e231](https://github.com/MorpheApp/morphe-cli/commit/ea6e2313268a8111c8a21faf425bda1eda534f35))
+* Use latest patcher dev release ([2b33a6c](https://github.com/MorpheApp/morphe-cli/commit/2b33a6cc0523be29cb7dcd86990de7e0e08c0c87))
+* Use latest patcher dev release ([baf34de](https://github.com/MorpheApp/morphe-cli/commit/baf34de0857f1c94b044aedbac33485b51ab3f2c))
+* Use latest patcher dev release ([9cf07f9](https://github.com/MorpheApp/morphe-cli/commit/9cf07f922fb5129aeeee2de0a3e990fecadae4fb))
+* Use latest patcher dev release ([791257c](https://github.com/MorpheApp/morphe-cli/commit/791257c48983c1a006eca87abd458331141252f1))
+* Use latest patcher dev release ([7cc195a](https://github.com/MorpheApp/morphe-cli/commit/7cc195a59d37521751fcbe5f15d267b3efbeeb0a))
+* Use latest patcher dev release ([f8bd235](https://github.com/MorpheApp/morphe-cli/commit/f8bd2354438dd3a1e969610d43dca224e9d3ef63))
+
+
+### Features
+
+* Add `--out` and `--patches` arguments to `list-patches` command ([#60](https://github.com/MorpheApp/morphe-cli/issues/60)) ([ef9fc48](https://github.com/MorpheApp/morphe-cli/commit/ef9fc482fcc6de2e51741b23ce12729991d6b36d))
+* Add ARSCLib support ([#55](https://github.com/MorpheApp/morphe-cli/issues/55)) ([07c3f7e](https://github.com/MorpheApp/morphe-cli/commit/07c3f7ec50d52739ee2695f52c3c7182f2287ecf))
+
+# [1.5.0-dev.11](https://github.com/MorpheApp/morphe-cli/compare/v1.5.0-dev.10...v1.5.0-dev.11) (2026-03-05)
+
+
+### Bug Fixes
+
+* Use latest patcher dev release ([ea6e231](https://github.com/MorpheApp/morphe-cli/commit/ea6e2313268a8111c8a21faf425bda1eda534f35))
+
+# [1.5.0-dev.10](https://github.com/MorpheApp/morphe-cli/compare/v1.5.0-dev.9...v1.5.0-dev.10) (2026-03-05)
+
+
+### Bug Fixes
+
+* Use latest patcher dev release ([2b33a6c](https://github.com/MorpheApp/morphe-cli/commit/2b33a6cc0523be29cb7dcd86990de7e0e08c0c87))
+
+# [1.5.0-dev.9](https://github.com/MorpheApp/morphe-cli/compare/v1.5.0-dev.8...v1.5.0-dev.9) (2026-03-02)
+
+
+### Bug Fixes
+
+* Use latest patcher dev release ([baf34de](https://github.com/MorpheApp/morphe-cli/commit/baf34de0857f1c94b044aedbac33485b51ab3f2c))
+
+# [1.5.0-dev.8](https://github.com/MorpheApp/morphe-cli/compare/v1.5.0-dev.7...v1.5.0-dev.8) (2026-03-02)
+
+
+### Bug Fixes
+
+* Use latest patcher dev release ([9cf07f9](https://github.com/MorpheApp/morphe-cli/commit/9cf07f922fb5129aeeee2de0a3e990fecadae4fb))
+
 # [1.5.0-dev.7](https://github.com/MorpheApp/morphe-cli/compare/v1.5.0-dev.6...v1.5.0-dev.7) (2026-02-28)